// src/pages/Game/FishGame/types.d.ts
export type ScreenType = 'welcome' | 'playing' | 'summary'

export type ImageResourcesType = {
  bg_1: string
  bg_2: string
  btn: string
  btn_answer: string
  btn_answer_off: string
  btn_back: string
  btn_null: string
  btn_sound_off: string
  btn_sound_on: string
  btn_summary: string
  game_name: string
  ico_board_correct: string
  ico_board_summary: string
  ico_board_text: string
  ico_fish_1: string
  ico_fish_2: string
  ico_fish_3: string
  ico_fish_4: string
  ico_fish_5: string
  ico_fish_6: string
  ico_fish_7: string
  ico_fish_8: string
  ico_fish_9: string
  ico_fish_10: string
  ico_heart_empty: string
  ico_heart_full: string
  ico_result_correct: string
  ico_result_wrong: string
  ico_robo_happy: string
  ico_robo_sad: string
  ico_shark_fat: string
  ico_shark_hungry: string
  ico_shark_hungry_extra: string
  ico_sound: string
  ico_scallop: string
  logo: string
}
export type AudioResourcesType = {
  sound_amazing: string
  sound_break: string
  sound_correct: string
  sound_ohno: string
  sound_wrong: string
  theme: string
  theme_bubble: string
  theme_summary: string
}